HW1 1 Finished Opt started

+++ src/main/java/ru/javawebinar/topjava/model/Meal.java
@@ -5,7 +5,7 @@ import java.time.LocalDateTime;
 import java.time.LocalTime;
 
 public class Meal {
-    private long id;
+    private Integer id;
 
     private final LocalDateTime dateTime;
 
@@ -19,18 +19,18 @@ public class Meal {
         this.calories = calories;
     }
 
-    public Meal(long id, LocalDateTime dateTime, String description, int calories) {
+    public Meal(Integer id, LocalDateTime dateTime, String description, int calories) {
         this.id = id;
         this.dateTime = dateTime;
         this.description = description;
         this.calories = calories;
     }
 
-    public long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
