+++ src/main/java/ru/javawebinar/topjava/model/MealTo.java
@@ -3,7 +3,7 @@ package ru.javawebinar.topjava.model;
 import java.time.LocalDateTime;
 
 public class MealTo {
-    private final long id;
+    private final Integer id;
 
     private final LocalDateTime dateTime;
 
@@ -13,7 +13,7 @@ public class MealTo {
 
     private final boolean excess;
 
-    public MealTo(long id, LocalDateTime dateTime, String description, int calories, boolean excess) {
+    public MealTo(Integer id, LocalDateTime dateTime, String description, int calories, boolean excess) {
         this.id = id;
         this.dateTime = dateTime;
         this.description = description;
@@ -39,7 +39,7 @@ public class MealTo {
                 '}';
     }
 
-    public long getId() {
+    public Integer getId() {
         return id;
     }
